Fix Pocket API working outside of player inventory
This makes Pocket API not equip/unequip upgrades when the pocket computer is
outside of the player inventory (e.g. dragging, dropped, placed in a chest).

When a pocket computer is outside of player inventory, it does not receive
Item.onUpdate event, and does not call ServerComputer.keepAlive(),
and so its ServerComputer.m_ticksSincePing increases by tick.
This detects it by !( computer.m_ticksSincePing == 0 ) and throws LuaException.